requirements.txt
pyjwt~=2.1
pytest==7.1.3
requests==2.31.0
cryptography>=39
sigstore-protobuf-specs~=0.2.0
